Split ui states.
