+++ src/ui/Ui/src/StateNull.cpp
@@ -1,0 +1,8 @@
+#include "StateNull.h"
+
+using namespace Ui;
+
+IUiState* StateNull::update (App::DataModel& /*model*/)
+{
+    return nullptr;
+}
